+++ src/sagas/ideas.js
@@ -1,9 +1,10 @@
 import { put, all, takeLatest, fork, call } from 'redux-saga/effects';
 import { actions, result, error } from '../components/Ideas/actions';
+import ideaService from '../services/idea';
 
 export const fetchIdeas = async () => {
     try {
-        const response = await fetch(`http://localhost:8082/ideas`);
+        const response = await ideaService.get();
         return await response.json();
     } catch (e) {
         catchError(e)